chore(components): rename legacy-ui/ to site/ and drop @/UI alias
The folder previously called legacy-ui/ wasn't legacy — it holds active
bespoke composition components (Header, Footer, Nav, AppLayoutPage,
Hero, Searchbar, BackToTop, EmptyResult, MobileNav, ThemeMenu, link
wrappers, image wrappers) that compose shadcn primitives plus
project-specific state. The legacy- prefix was a transitional concession
during the shadcn migration to dodge the Windows case-collision with
ui/. Rename to site/ now that the migration is settled.

Mechanical changes:
- git mv src/components/legacy-ui → src/components/site
- sed sweep: @/UI/ → @/components/site/, @/components/legacy-ui/ →
  @/components/site/ (~38 files)
- tsconfig.json: drop @/UI and @/UI/* path aliases (no longer needed)
- .prettierrc.js: drop the @/UI group from importOrder
- CLAUDE.md: update Path Aliases + Component Layout sections to
  describe ui/ + site/ instead of legacy-ui/

Verified locally: yarn type-check, yarn lint, yarn test (48/48),
yarn build (50 routes) — all clean.
