+++ CLAUDE.md
@@ -62,18 +62,17 @@ All blog, portfolio, and research entries are MDX files on disk — there's no C
 `tsconfig.json` baseUrl is `./src`. Use:
 
 - `@/*` → `src/*`
-- `@/UI` and `@/UI/*` → `src/components/legacy-ui` and `src/components/legacy-ui/*` (transitional alias for the bespoke site components — see "Component Layout" below)
 
-Imports are auto-sorted by `@ianvs/prettier-plugin-sort-imports` (order defined in `.prettierrc.js`): `@/components` → `@/UI` → `@/services` → `@/libs` → `@/...` → relative → third-party → `@/styles/*`. Empty strings in `importOrder` produce blank-line separators between groups. Don't reshuffle by hand; let Prettier do it.
+Imports are auto-sorted by `@ianvs/prettier-plugin-sort-imports` (order defined in `.prettierrc.js`): `@/components` → `@/services` → `@/libs` → `@/...` → relative → third-party → `@/styles/*`. Empty strings in `importOrder` produce blank-line separators between groups. Don't reshuffle by hand; let Prettier do it.
 
 ## Component Layout
 
 Two roots under `src/components/`:
 
 - **`ui/`** — shadcn/ui primitives. Every file here is generated by `npx shadcn@latest add <name>` and follows shadcn's data-slot, variant, and class conventions. Don't edit by hand unless intentionally extending; instead use `npx shadcn@latest add <name> --diff` to merge upstream changes. Built on `radix-ui` (configured in `components.json` as `style: radix-nova`); `radix-ui` is **not** a competing UI system — it's shadcn's primitive layer.
-- **`legacy-ui/`** — bespoke composition components specific to this site (Header, Footer, Nav, AppLayoutPage, Hero, Searchbar, BackToTop, EmptyResult, MobileNav, ThemeMenu, links, image wrappers). They compose `ui/` primitives + project state. The folder is named "legacy-ui" because it predates shadcn; the contents now use shadcn primitives internally and the rename is a future-PR concern.
+- **`site/`** — bespoke composition components specific to this site (Header, Footer, Nav, AppLayoutPage, Hero, Searchbar, BackToTop, EmptyResult, MobileNav, ThemeMenu, links, image wrappers). They compose `ui/` primitives + project state.
 
-When you need a new primitive, **search shadcn first** (`npx shadcn@latest search`). Only fall back to a bespoke component in `legacy-ui/` (or a feature folder) if no primitive fits.
+When you need a new primitive, **search shadcn first** (`npx shadcn@latest search`). Only fall back to a bespoke component in `site/` (or a feature folder) if no primitive fits.
 
 ## Theme System
