+++ src/app/blog/page.tsx
@@ -1,5 +1,5 @@
 import { BlogPageClient } from '@/components/blog/BlogPageClient'
-import { AppLayoutPage } from '@/components/legacy-ui/templates/AppLayoutPage'
+import { AppLayoutPage } from '@/components/site/templates/AppLayoutPage'
 
 import { getContents } from '@/services'
