+++ src/components/content/blog/AdjacentPosts.tsx
@@ -1,4 +1,4 @@
-import { UnstyledLink } from '@/UI/links'
+import { UnstyledLink } from '@/components/site/links'
 
 import type { AdjacentPosts as AdjacentPostsData } from '@/libs/sorters/getAdjacentPosts'
