+++ src/components/content/blog/AuthorSection.tsx
@@ -1,5 +1,5 @@
-import { WrappedImage } from '@/UI/images'
-import { UnderlineLink } from '@/UI/links'
+import { WrappedImage } from '@/components/site/images'
+import { UnderlineLink } from '@/components/site/links'
 
 import { useMemo } from 'react'
