+++ src/components/content/mdx/Headings.tsx
@@ -1,6 +1,6 @@
 'use client'
 
-import { UnstyledLink } from '@/UI/links'
+import { UnstyledLink } from '@/components/site/links'
 
 import { twclsx } from '@/libs'
