+++ src/components/content/portfolio/HeadingPortfolio.tsx
@@ -1,5 +1,5 @@
 import { CategoryLabel } from '@/components/content/portfolio/CategoryLabel'
-import { UnstyledLink } from '@/components/legacy-ui/links'
+import { UnstyledLink } from '@/components/site/links'
 
 import { twclsx } from '@/libs'
