+++ src/components/content/research/HeadingResearch.tsx
@@ -1,4 +1,4 @@
-import { UnstyledLink } from '@/components/legacy-ui/links'
+import { UnstyledLink } from '@/components/site/links'
 
 import { SITE_AUTHOR } from '@/libs/constants/site'
 import { twclsx } from '@/libs/twclsx'
