+++ src/components/site/images/WrappedImage.tsx
